requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
scipy>=1.11.0
pyyaml>=6.0
pillow>=10.0.0

# Scraping (mode real uniquement)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
